Move bundle-only test to Go-only suite
